+++ mangadex_downloader/format/comic_book.py
@@ -127,6 +127,8 @@ class ComicBookArchive(BaseFormat):
         manga = self.manga
         worker = self.create_worker()
 
+        self.write_tachiyomi_info()
+
         # Begin downloading
         for chap_class, images in manga.chapters.iter(**self.kwargs_iter):
             chap_name = chap_class.get_simplified_name()